requirements.txt
openai>=1.0.0
nltk>=3.8.1
textblob>=0.17.1
pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
